+++ src/pages/Admin/Charity/Widget/Configurer/index.ts
