+++ src/pages/Admin/Charity/Widget/Configurer/types.ts
