+++ src/pages/Admin/Charity/Widget/Preview/index.ts
