+++ src/pages/Admin/Charity/Widget/Snippet/createWidgerUrl.ts
@@ -1,13 +1,9 @@
 import { URL_PARAMS, UrlParamValues } from "pages/DonateWidget";
 import { isEmpty } from "helpers";
-import { DAPP_URL, IS_TEST } from "constants/env";
 import { appRoutes } from "constants/routes";
-import { FormValues } from "./schema";
-
-const APP_URL = IS_TEST ? "http://localhost:4200" : `${DAPP_URL}`;
 
 export default function createWidgetUrl(formValues: FormValues) {
-  const rootUrl = `${APP_URL}${appRoutes.donate_widget}/${formValues.endowment.id}?apiKey=API_KEY`;
+  const rootUrl = `${window.location.origin}/${appRoutes.donate_widget}/${formValues.endowment.id}?apiKey=API_KEY`;
 
   const param1 = append(formValues.hideText, URL_PARAMS.hideText);
   const param2 = append(
