+++ src/pages/Admin/Charity/Widget/Snippet/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Snippet";
