+++ src/pages/Admin/Charity/Widget/Widget.tsx
@@ -3,12 +3,13 @@ import Copier from "components/Copier";
 import { PAYMENT_WORDS } from "constants/common";
 import { adminRoutes } from "constants/routes";
 import Seo from "../Seo";
-import WidgetExample from "./WidgetExample";
-import WidgetUrlGenerator from "./WidgetUrlGenerator";
+import Configurer from "./Configurer";
+import Preview from "./Preview";
+import Snippet from "./Snippet";
 
 const TITLE_STYLE = "text-lg sm:text-2xl";
 
-export default function WidgetConfigurer() {
+export default function Widget() {
   //widget configurer is used in admin
   const { id } = useAdminContext();
   return (
@@ -35,31 +36,12 @@ export default function WidgetConfigurer() {
       <div className="grid xl:grid-cols-2 max-xl:justify-center gap-10">
         <section className="xl:order-2 flex flex-col gap-3 items-center xl:items-start justify-self-stretch">
           <h2 className={TITLE_STYLE}>Configure your widget</h2>
-          <WidgetUrlGenerator />
-          <WidgetSnippet />
+          <Configurer />
+          <Snippet />
         </section>
 
-        <section className="flex flex-col gap-3 max-xl:items-center">
-          <h2 className={TITLE_STYLE}>That's what our widget looks like:</h2>
-          <WidgetExample />
-        </section>
+        <Preview />
       </div>
     </div>
-  );
-}
-
-function WidgetSnippet() {
-  return (
-    <>
-      <h2 className={`${TITLE_STYLE} mt-10`}>
-        Copy / paste this code snippet:
-      </h2>
-      <div className="flex items-center justify-center gap-4 h-32 w-full max-w-xl px-10 rounded bg-gray-l3 dark:bg-blue-d4">
-        <span className="w-full text-sm sm:text-base font-mono break-all">
-          {""}
-        </span>
-        <Copier classes="w-10 h-10 hover:text-orange" text={""} />
-      </div>
-    </>
   );
 }
