+++ src/pages/Admin/Charity/Widget/constants.ts
@@ -1,0 +1,1 @@
+export const TITLE_STYLE = "text-lg sm:text-2xl";
