+++ src/pages/Admin/Charity/Widget/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Widget";
