+++ src/pages/Admin/Charity/Widget/useWidgetConfigurer.ts
@@ -2,9 +2,7 @@ import { useCallback, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useParams } from "react-router-dom";
 import { idParamToNum } from "helpers";
-import createWidgerUrl from "./createWidgerUrl";
-import { FormValues } from "./schema";
-import useLoadDefaultEndowmentName from "./useLoadDefaultEndowmentName";
+import createWidgerUrl from "./Snippet/createWidgerUrl";
 
 const NO_ID_MESSAGE = "Please select an organization";
