--- src/pages/Admin/Charity/WidgetConfigurer/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./WidgetConfigurer";
